requirements.txt
anyio==3.6.2
APScheduler==3.9.1.post1
certifi==2022.12.7
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
python-telegram-bot==20.0a6
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.44
SQLAlchemy-Utils==0.38.3
tzdata==2022.7
tzlocal==4.2
